Consolidate content item wrappers, disable generic wrappers for now
